pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["imaanick <nickolaytrusov01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "==2.*"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pytest = "^8.3.3"
httpx = "^0.27.2"
setuptools = "^75.2.0"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
